+++ examples/ErrorExample.hs
@@ -7,7 +7,6 @@ module Main where
 
 import qualified Knit.Report as K    
 
-import qualified Polysemy.Error                as PE
 import qualified Data.Map                      as M
 import qualified Data.Text.IO                  as T
 import qualified Data.Text.Lazy                as TL
@@ -23,8 +22,6 @@ templateVars = M.fromList
 --  , ("tufte","True")
   ]
 
-newtype MyError = MyError T.Text
-
 main :: IO ()
 main = do
   let pandocWriterConfig = K.PandocWriterConfig (Just "pandoc-templates/minWithVega-pandoc.html")  templateVars K.mindocOptionsF
@@ -35,16 +32,7 @@ main = do
         $ TL.toStrict
         $ htmlAsText
     Left err -> putStrLn $ "Pandoc Error: " ++ show err
-  resCustomE <- K.knitHtml (Just "SimpleExample.Main") K.logAll pandocWriterConfig (PE.runError @MyError $ makeDocWithKnitError)
-  case resCustomE of
-    Right resPandocE -> case resPandocE of
-      Right htmlAsText ->
-        T.writeFile "examples/html/example_simple.html"
-                        $ TL.toStrict
-                        $ htmlAsText
-      Left err -> putStrLn $ "Pandoc Error: " ++ show err
-    Left (MyError msg) -> putStrLn $ "Custom Error: " ++ show msg 
-
+    
 md1 :: T.Text
 md1 = [here|
 ## Some example markdown
@@ -66,24 +54,6 @@ makeDocWithKnitError = K.wrapPrefix "makeDocWithKnitError" $ do
   K.addLatex "Overused favorite equation: $e^{i\\pi} + 1 = 0$"
   K.logLE K.Info "adding a visualization..."
   K.knitError "Uh oh!  Something went wrong which I am explaining with this message."
-  K.addMarkDown "## An example hvega visualization"
-  K.addHvega "someID" exampleVis
-
-
-
-makeDocWithExtraError :: (K.Member K.ToPandoc effs -- required for the single-document variant
-                         , K.PandocEffects effs -- all effects for knitting
-                         , K.Member (PE.Error MyError) effs
-                         ) 
-                      => K.Sem effs ()
-makeDocWithExtraError = K.wrapPrefix "makeDocWithKnitError" $ do
-  K.logLE K.Info "adding some markdown..."
-  K.addMarkDown md1
-  K.logLE K.Info "adding some latex..."
-  K.addMarkDown "## Some example latex"
-  K.addLatex "Overused favorite equation: $e^{i\\pi} + 1 = 0$"
-  K.logLE K.Info "adding a visualization..."
-  PE.throw @MyError "A custom error with a message"
   K.addMarkDown "## An example hvega visualization"
   K.addHvega "someID" exampleVis